requirements.txt
Django==5.1.6
psycopg2-binary==2.9.10
python-decouple==3.8
pillow==11.1.0
django-countries==7.6.1
pycountry==24.6.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
drf-yasg==1.21.8
drf-nested-routers==0.94.1
djangorestframework-csv==3.0.2
django-cors-headers==4.6.0
django-phonenumber-field[phonenumberslite]==8.0.0
phonenumbers==8.13.55
django-money==3.5.4
celery==5.4.0
requests==2.32.3
pytest==8.3.5
pytest-django==4.10.0
pydantic==2.10.6
django-crontab==0.7.1
django-celery-results==2.6.0
redis==5.2.1
pyfiglet
termcolor
pandas
django-simple-history
telethon>=1.29.0
channels==4.0.0
channels-redis==4.2.0
daphne
django-extensions
coolname
python-magic==0.4.27
django-storages==1.14.4
boto3==1.35.0
google-cloud-storage==2.18.0
django-silk
django-encrypted-model-fields==0.6.5
reportlab>=4.0
langgraph>=0.2.0
langgraph-checkpoint-postgres
